docs: clarify uv_cwd troubleshooting
