+++ package.json
@@ -1,6 +1,15 @@
 {
   "name": "graphing-app",
   "version": "1.0.0",
-
-
+  "description": "Simple algebra graphing app",
+  "license": "MIT",
+  "type": "module",
+  "scripts": {
+    "test": "node --test",
+    "start": "http-server ."
+  },
+  "devDependencies": {
+    "http-server": "^14.1.1"
+  }
 }
+
