+++ tests/toFraction.test.js
@@ -1,11 +1,21 @@
-const { toFraction } = require('../app');
+import { readFileSync } from 'fs';
+import { runInNewContext } from 'vm';
+import test from 'node:test';
+import assert from 'node:assert';
 
-describe('toFraction', () => {
-  test('converts decimal to simplest fraction', () => {
-    expect(toFraction(0.5)).toEqual({ numerator: 1, denominator: 2 });
-  });
+const code = readFileSync(new URL('../app.js', import.meta.url));
+const sandbox = {
+  document: { getElementById: () => ({ addEventListener: () => {} }) },
+  window: {},
+  MathJax: {},
+  Plotly: { newPlot: () => {} }
+};
+runInNewContext(code, sandbox);
 
-  test('reduces fractions correctly', () => {
-    expect(toFraction(2.75)).toEqual({ numerator: 11, denominator: 4 });
-  });
+const { toFraction } = sandbox;
+
+test('toFraction reduces numbers', () => {
+  assert.strictEqual(toFraction(0.75), '3/4');
+  assert.strictEqual(toFraction(-2.5), '-5/2');
 });
+
